pyproject.toml
[project]
name = "orchestor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.5",
    "google-adk>=1.25.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.10",
    "langchain-mcp-adapters>=0.2.1",
    "matplotlib>=3.10.8",
    "pydantic>=2.12.5",
]
